+++ apps/site/src/components/page-sidebar.tsx
@@ -58,7 +58,7 @@ type SidebarPageSectionProps = {
 
 const highlightSection = (isSectionSelected: boolean) => ({
   borderLeft: `3px solid ${
-    isSectionSelected ? themeImport.palette.purple[700] : "white"
+    isSectionSelected ? themeImport.palette.purple[700] : "transparent"
   }`,
 });
 
@@ -98,7 +98,9 @@ const SidebarPageSection: FunctionComponent<SidebarPageSectionProps> = ({
         justifyContent="space-between"
         visibility="inherit"
         bgcolor={
-          isSectionSelected ? (theme) => theme.palette.purple[100] : "white"
+          isSectionSelected
+            ? (theme) => theme.palette.purple[100]
+            : "transparent"
         }
         pr={1}
       >
@@ -191,7 +193,7 @@ const SidebarPage: FunctionComponent<SidebarPageProps> = ({
   const { asPath } = router;
   const pathWithoutParams = generatePathWithoutParams(asPath);
 
-  const { href, title, sections, subPages } = page;
+  const { href, title, sections = [], subPages = [] } = page;
 
   const isSelected =
     pathWithoutParams === href || pathWithoutParams === `${href}#`;
@@ -205,7 +207,9 @@ const SidebarPage: FunctionComponent<SidebarPageProps> = ({
         display="flex"
         alignItems="center"
         justifyContent="space-between"
-        bgcolor={isSelected ? (theme) => theme.palette.purple[100] : "white"}
+        bgcolor={
+          isSelected ? (theme) => theme.palette.purple[100] : "transparent"
+        }
         pr={1}
       >
         <SidebarLink
@@ -331,14 +335,18 @@ const getInitialOpenedPages = (params: {
     if (pathWithoutParams === href || pathWithoutParams === `${href}#`) {
       return [href];
     }
-    const sectionPath = findSectionPath(href, sections, pathWithoutParams);
+    const sectionPath = findSectionPath(
+      href,
+      sections ?? [],
+      pathWithoutParams,
+    );
 
     if (sectionPath) {
       return [href, ...sectionPath];
     }
 
     const openSubPages = getInitialOpenedPages({
-      pages: subPages,
+      pages: subPages ?? [],
       asPath: pathWithoutParams,
     });
 
@@ -447,7 +455,7 @@ export const Sidebar: FunctionComponent<SidebarProps> = ({
         ) : pages.length === 1 ? (
           <>
             {/* When the sidebar is only displaying one page, we can display its sub-sections and sub-pages directly */}
-            {pages[0]!.sections.map((section, i) => (
+            {pages[0]!.sections?.map((section, i) => (
               <SidebarPageSection
                 key={section.anchor}
                 isSelectedByDefault={i === 0}
@@ -459,7 +467,7 @@ export const Sidebar: FunctionComponent<SidebarProps> = ({
                 setOpenedPages={setOpenedPages}
               />
             ))}
-            {pages[0]!.subPages.map((subpage) => (
+            {pages[0]!.subPages?.map((subpage) => (
               <SidebarPage
                 key={subpage.href}
                 depth={0}