+++ apps/site/src/pages/_document.page.tsx
@@ -15,6 +15,7 @@ class MyDocument extends Document {
             href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700;800;900&display=swap"
             rel="stylesheet"
           />
+          <link rel="stylesheet" href="https://use.typekit.net/muv5tib.css" />
           <link
             rel="apple-touch-icon"
             sizes="180x180"