+++ apps/site/src/theme/typography.ts
@@ -4,60 +4,61 @@ import { customColors } from "./palette";
 
 const defaultTheme = createTheme();
 
-const fallbackFonts = [`"Helvetica"`, `"Arial"`, "sans-serif"];
+export const fallbackFonts = [`"Helvetica"`, `"Arial"`, "sans-serif"];
 
+export const HEADING_FONT_FAMILY = `"colfax-web"`;
+export const COPY_FONT_FAMILY = `"Inter"`;
 export const CODE_FONT_FAMILY = `"JetBrains Mono", Monaco, monospace`;
 
 export const typography: ThemeOptions["typography"] = {
-  fontFamily: [`"Inter"`, ...fallbackFonts].join(", "),
   fontSize: 16,
   bpTitle: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-6)",
     lineHeight: 1,
     fontWeight: 700,
     color: customColors.gray[90],
   },
   bpSubtitle: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-3)",
     lineHeight: 1.1,
     fontWeight: 200,
     color: customColors.gray[90],
   },
   bpHeading1: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-5)",
     lineHeight: 1.1,
     fontWeight: 700,
     color: customColors.gray[90],
   },
   bpHeading2: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-4)",
     lineHeight: 1.2,
     color: customColors.gray[90],
   },
   bpHeading3: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-3)",
     lineHeight: 1.1,
     color: customColors.gray[80],
   },
   bpHeading4: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-2)",
     lineHeight: 1.1,
     color: customColors.gray[80],
   },
   bpHeading5: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-1)",
     lineHeight: 1.1,
     color: customColors.gray[80],
   },
   bpSmallCaps: {
-    fontFamily: [`"Apercu Pro"`, ...fallbackFonts].join(", "),
+    fontFamily: [HEADING_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step--1)",
     lineHeight: 1.3,
     color: customColors.gray[80],
@@ -65,11 +66,13 @@ export const typography: ThemeOptions["typography"] = {
     letterSpacing: "0.05em",
   },
   bpLargeText: {
+    fontFamily: [COPY_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-1)",
     lineHeight: 1.1,
     color: customColors.gray[90],
   },
   bpBodyCopy: {
+    fontFamily: [COPY_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontSize: "var(--step-0)",
     fontWeight: 400,
     lineHeight: 1.7,
@@ -86,12 +89,14 @@ export const typography: ThemeOptions["typography"] = {
     /** @todo: figure out how to type this */
   } as any,
   bpSmallCopy: {
+    fontFamily: [COPY_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontWeight: 500,
     fontSize: "var(--step--1)",
     lineHeight: 1.5,
     color: customColors.gray[90],
   },
   bpMicroCopy: {
+    fontFamily: [COPY_FONT_FAMILY, ...fallbackFonts].join(", "),
     fontWeight: 500,
     fontSize: "var(--step--2)",
     lineHeight: 1.1,