+++ apps/site/tests/integration/page-footer.test.ts
@@ -9,45 +9,74 @@ test("page footer navigation works", async ({ page, browserName }) => {
 
   await page.locator("footer").scrollIntoViewIfNeeded();
 
+  // Socials
   const socialLinksSelector = page.locator(
     "[data-testid='footer-social-links']",
   );
   await expect(socialLinksSelector.locator("a").first()).toHaveAttribute(
     "href",
-    "https://twitter.com/blockprotocol",
+    "https://github.com/blockprotocol/blockprotocol",
   );
   await expect(socialLinksSelector.locator("a").nth(1)).toHaveAttribute(
     "href",
-    "/discord",
+    "https://twitter.com/blockprotocol",
   );
   await expect(socialLinksSelector.locator("a").nth(2)).toHaveAttribute(
     "href",
-    "https://github.com/blockprotocol/blockprotocol",
+    "/discord",
   );
   await expect(
     socialLinksSelector.locator("text=Star us on Github"),
   ).toHaveAttribute("href", "https://github.com/blockprotocol/blockprotocol");
 
+  // Learn more section
+  await expect(
+    page.locator("footer >> a:has-text('Documentation')"),
+  ).toHaveAttribute("href", "/docs");
+  await expect(
+    page.locator("footer >> a:has-text('Specification')"),
+  ).toHaveAttribute("href", "/docs/spec");
+  // await expect(page.locator("footer >> a:has-text('Pricing')")).toHaveAttribute(
+  //   "href",
+  //   "/pricing",
+  // );
   await expect(
     page.locator("footer >> a:has-text('Contact Us')"),
   ).toHaveAttribute("href", "/contact");
 
   await page.locator("footer").scrollIntoViewIfNeeded();
 
-  await expect(page.locator("footer >> a:text-is('Hub')")).toHaveAttribute(
+  // Discover section
+  await expect(
+    page.locator("footer >> a:text-is('Open-source blocks')"),
+  ).toHaveAttribute("href", "/hub");
+  await expect(
+    page.locator("footer >> a:text-is('Semantic types')"),
+  ).toHaveAttribute("href", "/hub");
+  await expect(
+    page.locator("footer >> a:text-is('API endpoints')"),
+  ).toHaveAttribute("href", "/hub");
+
+  // Publish section
+  await expect(page.locator("footer >> a:text-is('a block')")).toHaveAttribute(
     "href",
-    "/hub",
+    "/docs/developing-blocks#publish",
   );
+  // await expect(page.locator("footer >> a:text-is('a type')")).toHaveAttribute(
+  //   "href",
+  //   "/",
+  // );
+  // await expect(
+  //   page.locator("footer >> a:text-is('an endpoint')"),
+  // ).toHaveAttribute("href", "/");
 
-  await expect(
-    page.locator("footer >> a:has-text('Documentation')"),
-  ).toHaveAttribute("href", "/docs");
-
-  await expect(
-    page.locator("footer >> a:has-text('Specification')"),
-  ).toHaveAttribute("href", "/docs/spec");
-
-  await expect(
-    page.locator("footer >> a:has-text('Publish a Block')"),
-  ).toHaveAttribute("href", "/docs/developing-blocks#publish");
+  // Legal section
+  // await expect(page.locator("footer >> a:text-is('Terms')")).toHaveAttribute(
+  //   "href",
+  //   "/legal/terms",
+  // );
+  // await expect(page.locator("footer >> a:text-is('Privacy')")).toHaveAttribute(
+  //   "href",
+  //   "/legal/privacy",
+  // );
 });