+++ libs/block-scripts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # block-scripts
 
+## 0.0.19
+
+### Patch Changes
+
+- [#816](https://github.com/blockprotocol/blockprotocol/pull/816) [`380b1493`](https://github.com/blockprotocol/blockprotocol/commit/380b149326450f4cf9b8300182eb199aa8f6a62f) Thanks [@kachkaev](https://github.com/kachkaev)! - Remove unused dependencies
+
+- [#875](https://github.com/blockprotocol/blockprotocol/pull/875) [`a1ba0d7d`](https://github.com/blockprotocol/blockprotocol/commit/a1ba0d7d17971ee30586a673ce3d4f5bee6e65d1) Thanks [@kachkaev](https://github.com/kachkaev)! - Update package metadata (change path to source following repo restructuring)
+
 ## 0.0.18
 
 ### Patch Changes